+++ loader/src/main/java/net/neoforged/fml/loading/moddiscovery/locators/ISystemModSource.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-only
  */
 
-package net.neoforged.fml.loading.moddiscovery.providers;
+package net.neoforged.fml.loading.moddiscovery.locators;
 
 /**
  * Only mod file locators that extend from this interface can return mod files that may contain system mods.
