+++ loader/src/main/java/net/neoforged/fml/loading/moddiscovery/locators/MinecraftModInfo.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-only
  */
 
-package net.neoforged.fml.loading.moddiscovery.providers;
+package net.neoforged.fml.loading.moddiscovery.locators;
 
 import com.electronwill.nightconfig.core.Config;
 import java.util.List;
