+++ loader/src/main/java/net/neoforged/fml/loading/targets/CommonServerLaunchHandler.java
@@ -10,7 +10,7 @@ import java.util.function.Consumer;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.fml.loading.MavenCoordinate;
 import net.neoforged.fml.loading.VersionInfo;
-import net.neoforged.fml.loading.moddiscovery.providers.ProductionServerProvider;
+import net.neoforged.fml.loading.moddiscovery.locators.ProductionServerProvider;
 import net.neoforged.neoforgespi.locating.IModFileCandidateLocator;
 
 /**
